Create areaOfCircleUsingClass.cpp
This code defines a class called Circle that has a private attribute called radius and several public methods for interacting with the radius attribute. The Circle class has a constructor that takes an integer argument r, which is used to initialize the radius attribute. There are also public setRadius and getRadius methods for setting and accessing the radius attribute, respectively.

The code also defines a class called Utility with a public method called calculateArea. This method takes an object of the Circle class as an argument and uses the getRadius method of the Circle object to obtain the value of the radius attribute. It then calculates the area of a circle with this radius using the formula 3.14 * r * r and returns the result.

In the main function, the code prompts the user to enter a radius, creates an object of the Circle class called c1 using this radius, creates an object of the Utility class called u1, and calls the calculateArea method of u1 with c1 as an argument. The returned result is then printed to the console.