Create arrManipulation.cpp
This C++ program allows the user to perform various operations on an array, such as inserting an element, deleting an element, displaying all the elements in the array, printing the last element in the array, or exiting the program.

The program begins by declaring an array arr of size 100, and variables sizearr, ext, x, and i. The user is then prompted to enter the size of the array, and the elements of the array are stored using the storeArr function. The program enters a while loop, which continues until the value of ext becomes 0. The user is presented with a menu of options and is asked to enter their choice.

Based on the user's choice, the program performs the following operations:

If the user chooses to insert an element, the program first prints the current elements in the array using the printArr function. The user is then asked to enter the index at which they want to insert the element and the value of the element. The program then shifts all the elements after the specified index to the right by one position, and inserts the new element at the specified index. The modified array is then printed using the printArr function.

If the user chooses to delete an element, the program first prints the current elements in the array using the printArr function. The user is then asked to enter the index of the element they want to delete. The program then shifts all the elements after the specified index to the left by one position, effectively deleting the element at the specified index. The modified array is then printed using the printArr function.

If the user chooses to display all the elements in the array, the program simply prints the current elements in the array using the printArr function.

If the user chooses to print the last element in the array, the program simply prints the last element in the array.

If the user chooses to exit the program, the value of ext is set to 0, which terminates the while loop and the program.

After performing the chosen operation, the program prompts the user to either continue or exit the program. If the user chooses to continue, the while loop continues and the program presents the menu of options again. If the user chooses to exit, the program prints a message indicating that the program has been successfully exited.