Create findingKey.cpp
This C++ program creates a binary search tree (BST) and allows the user to insert and search for nodes in the tree.

The program begins by including the iostream library and using the std namespace.

The program then defines a Node class that has three members: data, left, and right. data is an int that stores the value of the node, and left and right are pointers to other Node objects that represent the left and right children of the node, respectively.

The program then defines a BST class that has two members: root and a constructor. root is a pointer to a Node object that represents the root of the BST, and the constructor initializes root to NULL. The BST class also has two member functions: insert and search.

The insert function takes as input a pointer to a Node object and an int value, and it returns a pointer to a Node object. It creates a new Node object with the given int value and sets its left and right children to NULL. If the input Node pointer is NULL, the function sets the input Node pointer to the new Node object. Otherwise, it compares the value of the new Node object with the value of the input Node object. If the value of the new Node object is greater than or equal to the value of the input Node object, the function recursively calls itself with the right child of the input Node object and the new Node object. If the value of the new Node object is less than the value of the input Node object, the function recursively calls itself with the left child of the input Node object and the new Node object. Finally, the function returns the input Node object.

The search function takes as input an int value and returns an int. It initializes a Node pointer temp to the root of the BST, and it enters a loop that continues until temp is NULL. Inside the loop, the function compares the value of the temp node with the input int value. If they are equal, the function returns the input int value. Otherwise, if the value of the temp node is greater than the input int value, the function sets temp to the left child of the temp node. If the value of the temp node is less than the input int value, the function sets temp to the right child of the temp node. If the loop ends without finding a matching node, the function returns -1.

In the main function, the program creates an instance of the BST class called t and declares some variables. It then prompts the user to enter an int value, which it stores in the num variable. It then calls the insert function on the root of the t object and the num variable, and assigns the returned value to the root of the t object. The program then enters a loop that prompts the user to enter another int value. If the user enters "y", the program reads in the int value and calls the insert function on the root of the t object and the value. If the user enters anything other than "y", the program is terminated and user is asked to enter the search value.