Create starGrid.cpp
The code begins with the inclusion of the bits/stdc++.h header file, which includes a set of standard C++ libraries.

The code then enters the main function, which is the entry point of the program.

The main function declares a variable t of type int, which is used to store the number of test cases. This value is read from the input stream using cin.

The main function then enters a while loop that will execute t number of times. Inside the loop, another variable n of type int is declared and used to store the size of the square. This value is also read from the input stream using cin.

Inside the loop, an object ob of the Solution class is created. The Solution class has a function called printSquare, which takes an integer n as input and prints a square of size n to the output stream. This function is called on the ob object, passing in the size of the square as an argument.

After all test cases have been processed, the main function returns 0, indicating that the program has completed execution successfully.

The Solution class has a single function, printSquare, which takes an integer n as input and prints a square of size n to the output stream. The function first checks if the value of n is within the valid range of 1 to 20. If it is, the function enters a loop that iterates n number of times. For each iteration, another loop is entered that also iterates n number of times. Inside this inner loop, a star (*) and a space are printed to the output stream. After the inner loop finishes executing, a newline character (\n) is printed, which causes the cursor to move to the next line. This process is repeated until all rows of the square have been printed.