Create simpleStarPyramid.cpp
The first line #include <bits/stdc++.h> is a directive to the preprocessor to include a standard C++ library called bits/stdc++.h. This library includes several commonly used C++ functions, such as input/output functions, string manipulation functions, and math functions.

The line using namespace std; tells the compiler to use the std namespace, which contains the functions and variables defined in the standard C++ library.

The class Solution is defined. It contains a public function called printSquare, which takes an integer n as a parameter. This function will be used to print a square of asterisks of size n.

Inside the printSquare function, there is a check to ensure that the size of the square, n, is within the valid range of 1 to 20. If n is outside this range, the function will not do anything.

Inside the if-statement, there are two nested for loops. The outer loop iterates through the rows of the square, and the inner loop iterates through the columns of the square. The variable i is used to keep track of the current row, and the variable j is used to keep track of the current column.

Inside the inner loop, the function prints an asterisk followed by a space. This is done for every column in the current row.

After finishing a row, the function moves to the next line by printing a newline character.

The for loops continue until all rows and columns have been printed, at which point the function finishes.

In the main function, the variable t is used to store the number of test cases. This is read in from standard input using the cin function.

There is a while loop that continues until all test cases have been processed. Inside the loop, the variable n is used to store the size of the square for the current test case. This is also read in from standard input.

An object of the Solution class is created and the printSquare function is called with the size of the square as an argument.

The while loop continues until all test cases have been processed, at which point the main function returns 0 to indicate that the program has finished successfully.