Create 7_Pyramid.cpp
This C++ program will output a right-angled triangle made of asterisks (*) to the console. The height of the triangle will be equal to the number of rows n specified by the user.Note that the number of asterisks in each row is equal to (2*r)-1, where r is the current row number. The number of spaces at the beginning of each row is calculated as the difference between n and the current row number.