Create 9_diamond.cpp
The output of this code will be a diamond pattern of asterisks (*) with a height and width of n, where n is the input given to the printDiamond function.

The diamond pattern will be composed of n lines, with the top half of the diamond being made up of n rows of asterisks and the bottom half being made up of n-1 rows of asterisks. The number of asterisks in each row will depend on the row number, with the number of asterisks increasing as the row number increases in the top half of the diamond and decreasing as the row number increases in the bottom half.