Rename rotatedTriangle.cpp to 10_rotatedTriangle.cpp
This code defines a class Solution with a member function printTriangle that takes an integer n as an argument. The function first checks if n is within the range 1 to 20. If it is, the function will print a pattern of asterisks (*) in the form of two triangles. The first triangle is constructed by printing n rows, where the number of asterisks in each row is equal to the row number. The second triangle is constructed by printing n-1 rows, where the number of asterisks in each row is equal to the row number.