Create 13_alphabitTriangle.cpp
The output of this code will be a series of alphabetically ascending triangles with a height of n. The first triangle will have a height of 1 and will consist of a single character 'A'. The second triangle will have a height of 2 and will consist of two characters 'A' and 'B', and so on. For example, if n is 5, the output will be:
A
AB
ABC
ABCD
ABCDE