Create 15_alphabitPyramid.cpp
The "printTriangle" method takes in an integer "n" as an input and prints out a triangle of characters to the console. The triangle consists of "n" rows, with the first row having 1 character, the second row having 2 characters, and so on until the last row, which has "n" characters. The characters printed in each row are capital letters of the English alphabet, starting with 'A' in the first row and increasing by one for each subsequent row.

For example, if the input value of "n" is 5, the method will output the following triangle:
A
BB
CCC
DDDD
EEEEE