Create 16_alphabitPyramid.cpp
The code first checks if the input value of "n" is within the range of 1 to 20. If it is, it enters a loop that iterates "n" number of times, with the variable "row" representing the current row of the triangle being printed.

For each row, the code first prints a number of spaces equal to "n" minus the current row number. This aligns the triangles so that they have a symmetrical appearance.
For example, if the input value of "n" is 4, the output will be:
   A
  ABA
 ABCBA
ABCDCBA