Create 19_spaceDiamond.cpp
he code defines a class Solution with a method printTriangle which takes an integer n as input and prints a pattern in the form of a triangle.

The pattern will have 2*n rows, with the top half of the triangle having n rows and the bottom half having n rows. Each row will have n columns.

The top half of the triangle will be an inverted triangle, with the first row having n asterisks (*), the second row having n-1 asterisks, and so on. Between each pair of asterisks, there will be 2 spaces.

The bottom half of the triangle will be a triangle, with the first row having 1 asterisk, the second row having 2 asterisks, and so on. Between each pair of asterisks, there will be 2*(n-1) - 2*rowNumber spaces.
Input: 5

Output:
**********
****  ****
***    ***
**      **
*        *
*        *
**      **
***    ***
****  ****
**********