Create 20_psaceshipPattern.cpp
It prints n rows of the pattern. Each row consists of the following:
A number of asterisks equal to the row number.
A number of spaces equal to n - row.
Another number of asterisks equal to the row number.
It then prints n - 1 more rows of the pattern. Each row is the same as the rows printed in step 1, but with one less asterisk at the beginning and end of the row.
Input: 5

Output:
*        *
**      **
***    ***
****  ****
**********
****  ****
***    ***
**      **
*        *