Create 22_gridPattern.cpp
The output of this code will be a series of n x n grids where the value of each cell in the grid is determined by its distance from the nearest edge of the grid. The value in each cell is equal to n minus the minimum distance from that cell to the nearest edge of the grid. The number of grids printed will be equal to the value of the variable "t", which is read in from standard input at the beginning of the main function.
Input: 4

Output:
4 4 4 4 4 4 4
4 3 3 3 3 3 4
4 3 2 2 2 3 4
4 3 2 1 2 3 4
4 3 2 2 2 3 4
4 3 3 3 3 3 4
4 4 4 4 4 4 4