Create gfg_countDigits.cpp
The main function reads an integer t which is the number of test cases to be run. Then it runs a loop t times. In each iteration it reads an integer N and creates an object of class Solution and calls the function evenlyDivides(N) which returns the number of digits in N that evenly divide N and it prints the returned value.